requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
python-dotenv>=0.20
requests>=2.20
gunicorn # 本番環境用WSGIサーバー
psycopg2-binary # PostgreSQLドライバー (PaaSでの利用を推奨)
Flask-Migrate>=4.0 # DBスキーマ変更の管理に (強く推奨)
python-dateutil
jpholiday # 日本の祝日を取得するためのライブラリ
psycopg2-binary
Flask-WTF
wtforms-sqlalchemy
icalendar
cryptography
Flask-Limiter
Flask-Login

# Flask-Login>=0.6 # ログイン管理の補助に (Authlibや自前でも可) (任意)
# Flask-Uploads-Fork # ファイルアップロード処理に (代替ライブラリも検討可) (任意)
# Pillow # 画像処理に (ファイルタイプ検証などで必要になる場合あり) (任意)
